chore(#8260): migrate rapidpro e2e test (#8267)
Ticket #8260.

This commit covers the following tests:
- RapidPro SMS Gateway - Webapp Terminating messages
  ✓ should fail with no incoming key configured
  ✓ should fail with no authentication provided
  ✓ should fail with incorrect authentication
  ✓ should fail with malformed authentication
  ✓ should fail when message was not created
  ✓ should create messages
  ✓ should create reports

- RapidPro SMS Gateway - Webapp originating messages and state updates
  ✓ should not call RapidPro endpoint when credentials are not set
  ✓ should call endpoint with set credentials
  ✓ should set correct states from broadcast api
  ✓ should update the states correctly from messages api
  ✓ should not poll messages for docs that are in pending or scheduled states or lack gateway ref
  ✓ should keep updating until record is in final state
  ✓ should poll for state updates and consume queue
  ✓ should handle rate limiting when polling
