+++ tests/e2e/default-mobile/reports/delete.wdio-spec.js
@@ -17,20 +17,24 @@ describe('Delete Reports', () => {
 
   const today = moment();
   const reports = [
-    reportFactory.build(
-      {
-        form: 'P',
-        reported_date: moment([today.year(), today.month(), 1, 23, 30]).subtract(4, 'month').valueOf()
-      },
-      { patient, submitter: onlineUser.contact, fields: { lmp_date: 'Feb 3, 2022' } },
-    ),
-    reportFactory.build(
-      {
-        form: 'P',
-        reported_date: moment([today.year(), today.month(), 12, 10, 30]).subtract(1, 'month').valueOf()
-      },
-      { patient, submitter: onlineUser.contact, fields: { lmp_date: 'Feb 16, 2022' } },
-    ),
+    reportFactory
+      .report()
+      .build(
+        {
+          form: 'P',
+          reported_date: moment([today.year(), today.month(), 1, 23, 30]).subtract(4, 'month').valueOf()
+        },
+        { patient, submitter: onlineUser.contact, fields: { lmp_date: 'Feb 3, 2022' } },
+      ),
+    reportFactory
+      .report()
+      .build(
+        {
+          form: 'P',
+          reported_date: moment([today.year(), today.month(), 12, 10, 30]).subtract(1, 'month').valueOf()
+        },
+        { patient, submitter: onlineUser.contact, fields: { lmp_date: 'Feb 16, 2022' } },
+      ),
   ];
 
   const savedReportIds = [];
