+++ tests/e2e/default/more-options-menu/offline-user/all-permissions.wdio-spec.js
@@ -39,17 +39,19 @@ const patient = personFactory.build({
   _id: uuid(),
   parent: { _id: clinic._id, parent: { _id: health_center._id, parent: { _id: district_hospital._id }}}
 });
-const xmlReport = reportFactory.build({ form: 'home_visit', content_type: 'xml' }, { patient, submitter: contact });
+const xmlReport = reportFactory
+  .report()
+  .build(
+    { form: 'home_visit', content_type: 'xml' },
+    { patient, submitter: contact }
+  );
 
-const smsReport = reportFactory.build(
-  {
-    form: 'P',
-    patient_id: patient._id,
-  },
-  {
-    patient, submitter: offlineUser.contact, fields: { lmp_date: 'Feb 3, 2022', patient_id: patient._id},
-  },
-);
+const smsReport = reportFactory
+  .report()
+  .build(
+    { form: 'P', patient_id: patient._id, },
+    { patient, submitter: offlineUser.contact, fields: { lmp_date: 'Feb 3, 2022', patient_id: patient._id }, },
+  );
 
 describe('More Options Menu - Offline User', async () => {
   let xmlReportId;
