+++ tests/e2e/default/more-options-menu/online-user/permissions-enabled.wdio-spec.js
@@ -37,16 +37,18 @@ const patient = personFactory.build({
   parent: { _id: clinic._id, parent: { _id: health_center._id, parent: { _id: district_hospital._id }}}
 });
 
-const xmlReport = reportFactory.build({ form: 'home_visit', content_type: 'xml' }, { patient, submitter: contact });
-const smsReport = reportFactory.build(
-  {
-    form: 'P',
-    patient_id: patient._id,
-  },
-  {
-    patient, submitter: contact, fields: { lmp_date: 'Dec 3, 2022', patient_id: patient._id},
-  },
-);
+const xmlReport = reportFactory
+  .report()
+  .build(
+    { form: 'home_visit', content_type: 'xml' },
+    { patient, submitter: contact }
+  );
+const smsReport = reportFactory
+  .report()
+  .build(
+    { form: 'P', patient_id: patient._id, },
+    { patient, submitter: contact, fields: { lmp_date: 'Dec 3, 2022', patient_id: patient._id}, },
+  );
 
 describe('Online User', async () => {
