+++ tests/e2e/default/reports/breadcrumbs.wdio-spec.js
@@ -54,18 +54,20 @@ describe('Reports tab breadcrumbs', () => {
 
   const today = moment();
   const reports = [
-    reportFactory.build(
-      {
-        form: 'P',
-        reported_date: moment([today.year(), today.month(), 1, 23, 30]).subtract(4, 'month').valueOf(),
-        patient_id: 'patient1',
-      },
-      {
-        patient,
-        submitter: offlineUser.contact,
-        fields: { lmp_date: 'Feb 3, 2022', patient_id: 'patient1' },
-      },
-    ),
+    reportFactory
+      .report()
+      .build(
+        {
+          form: 'P',
+          reported_date: moment([today.year(), today.month(), 1, 23, 30]).subtract(4, 'month').valueOf(),
+          patient_id: 'patient1',
+        },
+        {
+          patient,
+          submitter: offlineUser.contact,
+          fields: { lmp_date: 'Feb 3, 2022', patient_id: 'patient1' },
+        },
+      ),
   ];
 
   before(async () => {
