+++ tests/utils/index.js
@@ -845,6 +845,20 @@ const dockerGateway = () => {
   }
 };
 
+const dockerPlatformName = () => {
+  try {
+    return JSON.parse(execSync(`docker version --format '{{json .Server.Platform.Name}}'`));
+  } catch (error) {
+    console.log('docker version failed. NOTE this error is not relevant if running outside of docker');
+    console.log(error.message);
+  }
+  return null;
+};
+
+const isDockerDesktop = () => {
+  return (dockerPlatformName() || '').includes('Docker Desktop');
+};
+
 const getDockerVersion = () => {
   try {
     const response = execSync('docker-compose -v').toString();
@@ -856,10 +870,22 @@ const getDockerVersion = () => {
   }
 };
 
-const hostURL = (port = 80) => {
+const getHostRoot = () => {
+  if (isDockerDesktop()) {
+    // Docker Desktop networking requires a special host name for connecting to host machine.
+    // https://docs.docker.com/desktop/networking/#i-want-to-connect-from-a-container-to-a-service-on-the-host
+    return 'host.docker.internal';
+  }
   const gateway = dockerGateway();
-  const host = gateway && gateway[0] && gateway[0].Gateway ? gateway[0].Gateway : 'localhost';
-  const url = new URL(`http://${host}`);
+  if (gateway && gateway[0] && gateway[0].Gateway) {
+    return gateway[0].Gateway;
+  }
+
+  return 'localhost';
+};
+
+const hostURL = (port = 80) => {
+  const url = new URL(`http://${getHostRoot()}`);
   url.port = port;
   return url.href;
 };
